# [UX] Support `--infra` option and deprecate `cloud/region/zone`