--- a/.cursor/rules/format.mdc
+++ b/.cursor/rules/format.mdc
@@ -1,7 +1,0 @@
----
-description: 
-globs: 
-alwaysApply: true
----
-Line length should not exceed 80 characters.
-Do not use backslashes (`\`) for line continuation.