--- a/examples/admin_policy/task.yaml
+++ b/examples/admin_policy/task.yaml
@@ -1,5 +1,5 @@
 resources:
-  cloud: aws
+  infra: aws
   cpus: 2
   labels:
     other_labels: test