… .cursor/rules/test.mdc
… docs/source/cloud-setup/cloud-permissions/aws.rst
… docs/source/cloud-setup/cloud-permissions/gcp.rst
… docs/source/cloud-setup/quota.rst
… docs/source/compute/gpus.rst
… docs/source/compute/show-gpus-all.txt
… docs/source/compute/show-gpus-h100-8.txt
… docs/source/examples/auto-failover.rst
… docs/source/examples/managed-jobs.rst
… docs/source/getting-started/installation.rst
… docs/source/getting-started/quickstart.rst
… docs/source/reference/api-server/api-server.rst
… docs/source/reference/async.rst
… docs/source/reference/auto-stop.rst
… docs/source/reference/config.rst
… docs/source/reference/kubernetes/kubernetes-deployment.rst
… docs/source/reference/kubernetes/kubernetes-getting-started.rst
… docs/source/reference/kubernetes/kubernetes-priorities.rst
… docs/source/reference/kubernetes/kubernetes-setup.rst
… docs/source/reference/kubernetes/kubernetes-troubleshooting.rst
… docs/source/reference/kubernetes/multi-kubernetes.rst
… docs/source/reference/yaml-spec.rst
… docs/source/reservations/existing-machines.rst
… docs/source/reservations/reservations.rst
… docs/source/running-jobs/many-jobs.rst
… docs/source/serving/sky-serve.rst
… docs/source/serving/spot-policy.rst
… docs/source/serving/update.rst
… examples/autogluon.yaml
… examples/aws_efa/nccl_efa.yaml