<!-- Describe the changes in this PR -->
This PR updates the UX for the output, and add support for infra

Also, fixes #5600

## Main changes
1. Support `--infra` and deprecate `--cloud`, `--region`, and `--zone`
2. Show `INFRA` in outputs instead of cloud/region/zone
3. Reorder/polishing table outputs

```
sky launch --infra aws/us-east-1                                                                                 
Considered resources (1 node):
------------------------------------------------------------------------------
 INFRA             INSTANCE      vCPUs   Mem(GB)   GPUS   COST ($)   CHOSEN   
------------------------------------------------------------------------------
 AWS (us-east-1)   m6i.2xlarge   8       32        -      0.38          ✔     
------------------------------------------------------------------------------
Launching a new cluster 'sky-472f-zhwu'. Proceed? [Y/n]:
```

```
sky launch --infra gcp                                                                                           ✭ ✱
Considered resources (1 node):
------------------------------------------------------------------------------------
 INFRA                 INSTANCE        vCPUs   Mem(GB)   GPUS   COST ($)   CHOSEN   
------------------------------------------------------------------------------------
 GCP (us-central1-a)   n2-standard-8   8       32        -      0.39          ✔     
------------------------------------------------------------------------------------
Launching a new cluster 'sky-c16f-zhwu'. Proceed? [Y/n]:
```

```
sky launch --cloud k8s                                                                                           ✭ ✱
The --cloud, --region, and --zone options are deprecated. Use --infra instead.
Considered resources (1 node):
-------------------------------------------------------------------------------------------
 INFRA                                   INSTANCE   vCPUs   Mem(GB)   GPUS   COST ($)   CHOSEN   
-------------------------------------------------------------------------------------------
 Kubernetes (gke_sky-dev...ypilotalpha)  -          2       2         -      0.00          ✔     
-------------------------------------------------------------------------------------------
Launching a new cluster 'sky-5d56-zhwu'. Proceed? [Y/n]:
```

## Optimizer table
Original
```
sky launch --gpus H100:8
Considered resources (1 node):
-------------------------------------------------------------------------------------------------------------------------------------
 CLOUD        INSTANCE            vCPUs   Mem(GB)   ACCELERATORS   REGION/ZONE            COST ($)   CHOSEN   
-------------------------------------------------------------------------------------------------------------------------------------
 Kubernetes   2CPU--2GB--8H100    2       8         H100:8         gke_sky-dev-xxx        0.00       ✔     
 RunPod       8x_H100_SECURE      128     640       H100:8         CA                     19.12               
 Lambda       gpu_8x_h100_sxm5    208     1800      H100:8         us-east-1              23.92               
 GCP          a3-highgpu-8g       208     1872      H100:8         us-central1-a          46.02               
 AWS          p5.48xlarge         192     2048      H100:8         us-east-1              98.32               
-------------------------------------------------------------------------------------------------------------------------------------
Launching a new cluster 'sky-b84f-zhwu'. Proceed? [Y/n]:
```
New
```
sky launch --gpus H100:8                                                                                                                   
Considered resources (1 node):
---------------------------------------------------------------------------------------------------------
 INFRA                                    INSTANCE           vCPUs   Mem(GB)   GPUS     COST ($)   CHOSEN   
---------------------------------------------------------------------------------------------------------
 Kubernetes (gke_sky-dev...ypilotalpha)   -                  2       8         H100:8   0.00          ✔     
 RunPod (CA)                              8x_H100_SECURE     128     640       H100:8   19.12               
 Lambda (us-east-1)                       gpu_8x_h100_sxm5   208     1800      H100:8   23.92               
 GCP (us-central1-a)                      a3-highgpu-8g      208     1872      H100:8   46.02               
 AWS (us-east-1)                          p5.48xlarge        192     2048      H100:8   98.32               
---------------------------------------------------------------------------------------------------------
Launching a new cluster 'sky-3d09-zhwu'. Proceed? [Y/n]:
```


## Status table
Original:
```
sky status -u                                                                                                                                              
Clusters
NAME                          USER  LAUNCHED     RESOURCES                                                                  STATUS   AUTOSTOP  COMMAND                       
test-normal                   zhwu  17 hrs ago   1x GCP(n2-standard-2)                                                      UP       -         sky launch -c test-normal...  
dashboard-workspace           zhwu  6 days ago   1x GCP(n2-standard-4, ports=['46580'])                                     INIT     -         sky launch -c dashboard-w...  
benchmark                     zhwu  2 weeks ago  1x Nebius(gpu-h200-sxm_8gpu-128vcpu-1600gb, mem=750+, {'H200': 8}, dis...  STOPPED  -         sky exec -c benchmark qwe...  
sky-jobs-controller-9ce1ce58  zhwu  14 hrs ago   1x GCP(n2-highmem-4, disk_size=50)                                         STOPPED  10m       sky jobs queue -r   
```

New:
```
sky status -u                                                                                                         
Clusters
NAME                          USER  INFRA                            RESOURCES                                      STATUS   AUTOSTOP  LAUNCHED     
sky-f4e8-zhwu                 zhwu  Kubernetes (gke_sky-dev-xxx...)  1x(gpus=H100:1, cpus=2, mem=8, ...)            UP       -         2 hrs ago    
test-normal                   zhwu  GCP (us-central1-a)              1x(cpus=2, mem=8, type=n2-standard-2, ...)     UP       -         21 hrs ago   
dashboard-workspace           zhwu  GCP (us-central1-a)              1x(cpus=4, mem=16, type=n2-standard-4, ...)    INIT     -         7 days ago   
benchmark                     zhwu  Nebius (eu-west1)                1x(gpus=H200:8, type=gpu-h200-sxm_8g..., ...)  STOPPED  -         2 weeks ago  
sky-jobs-controller-9ce1ce58  zhwu  GCP (us-central1-a)              1x(cpus=4, mem=32, type=n2-highmem-4, ...)     STOPPED  10m       17 hrs ago   
```

## `sky check`
Original
```
To enable a cloud, follow the hints above and rerun: sky check
If any problems remain, refer to detailed docs at: https://docs.skypilot.co/en/latest/getting-started/installation.html

🎉 Enabled clouds 🎉
  AWS [compute, storage]
  GCP [compute, storage]
  Kubernetes [compute]
    Active context: gke_sky-dev-xxx
  Lambda [compute]
  RunPod [compute]

Using SkyPilot API server: http://127.0.0.1:46580
```

New
```
To enable a cloud, follow the hints above and rerun: sky check
If any problems remain, refer to detailed docs at: https://docs.skypilot.co/en/latest/getting-started/installation.html

🎉 Enabled infra 🎉
  AWS [compute, storage]
  GCP [compute, storage]
  Kubernetes [compute]
    Active context: gke_sky-dev-xxx
  Lambda [compute]
  RunPod [compute]

Using SkyPilot API server: http://127.0.0.1:46580
```

## `sky show-gpus`

```
sky show-gpus --infra k8s/my-h100-cluster                                                                                                       
Kubernetes GPUs
Context: my-h100-cluster
GPU   REQUESTABLE_QTY_PER_NODE  UTILIZATION  
H100  1, 2, 4, 8                5 of 8 free  
H200  1, 2, 4, 8                8 of 8 free  
Kubernetes per-node GPU availability
CONTEXT          NODE                                GPU   UTILIZATION  
my-h100-cluster  gke-xxx-default-pool-ff931856-6uvd  -     0 of 0 free  
my-h100-cluster  gke-xxx-largecpu-05dae726-1usy      H100  5 of 8 free  
my-h100-cluster  gke-xxx-largecpu-05dae726-4rxa      H200  8 of 8 free
```

## Dashboard

Original
<img width="1707" alt="image" src="https://github.com/user-attachments/assets/d7f82129-ef9e-482c-9f04-7055e288b64e" />


New
<img width="1691" alt="image" src="https://github.com/user-attachments/assets/76716e8a-fbcf-4b98-a0be-de80f58b83a3" />


### Cluster
Original
![image](https://github.com/user-attachments/assets/7f0eeaef-fd01-48de-ae8b-10cc0b61d1cf)


New
<img width="1706" alt="image" src="https://github.com/user-attachments/assets/c325874c-4f5f-492d-9efa-26ce59cc1900" />
<img width="1709" alt="image" src="https://github.com/user-attachments/assets/6f7133b1-3eba-4e1f-bf3f-30952c224c20" />




## TODOs
- [x] Update `sky check`
- [x] Update `sky show-gpus`
- [x] Update smoke tests
- [x] Check ordered and any_of logic
- [x] Update docs with the new table
- [x] Test with an old jobs controller
- [x] Test with old YAML that has `cloud`, `region`, `zone` only


## Future TODOs

- [ ] Update the doc to infra centric instead of cloud/region centric (cc'ing @concretevitamin)

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: install pre-commit (auto-check on commit) or `bash format.sh`
- [ ] Any manual or new tests for this PR (please specify below)
  - [x] Old controller, `jobs queue`, `jobs logs`, `jobs cancel` without `sky jobs launch` first
- [ ] All smoke tests: `/smoke-test` (CI) or `pytest tests/test_smoke.py` (local)
  - [x] `/smoke-test --aws` [4b79915](https://github.com/skypilot-org/skypilot/pull/5602/commits/4b799157940b27b9f77b6308f2d6069c3403a0e7)
  - [x] `/smoke-test --gcp` [4b79915](https://github.com/skypilot-org/skypilot/pull/5602/commits/4b799157940b27b9f77b6308f2d6069c3403a0e7) (except for `test_tpu`)
  - [x] `/smoke-test --kubernetes` [583360f](https://github.com/skypilot-org/skypilot/pull/5602/commits/583360f9670b667bd32795bf7d705776d5ae4901)
  - [x] `/smoke-test --azure` [be75cbe](https://github.com/skypilot-org/skypilot/pull/5602/commits/be75cbe33f8b6cf9c21e58ed8e8d757b0e829aa4)
- [ ] Relevant individual tests: `/smoke-test -k test_name` (CI) or `pytest tests/test_smoke.py::test_name` (local)
- [x] Backward compatibility: `/quicktest-core` (CI) or `pytest tests/smoke_tests/test_backward_compat.py` (local)
  - [x] `/quicktest-core` passed on [e39e339](https://github.com/skypilot-org/skypilot/pull/5602/commits/e39e3396de60a4fe6b1aaf003b61c67fec2b69f7)

<!-- CI commands (/-prefixed) can only be triggered by repo members -->