@Michaelvll 2025-05-17
/smoke-test --gcp