@Michaelvll 2025-05-18
/smoke-test --aws