@Michaelvll 2025-05-19
> I am a bit worried about updating the latest docs until this is in a stable release.

Good point! Adding a banner to address this issue in #5629