@Michaelvll 2025-05-20
/smoke-test --gcp