@zpoint 2025-05-20
https://buildkite.com/skypilot-1/release/settings/steps

We should update release pipeline as well